requirements.txt
# see https://pytorch.org/get-started/locally/
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu129

torch --index-url https://download.pytorch.org/whl/cu129
torchvision --index-url https://download.pytorch.org/whl/cu129

pymongo
dotenv
pandas
numpy
matplotlib
pyyaml
wandb
